feat(report): put revenue in the daily email, drop the heavy-user list
The daily report gathered "top 5 orgs by messages in 7 days" and no
revenue. Message volume turned out to track neither: the biggest storers
were often accounts that had never run a search, while the paying
customers sat mid-table.

Swaps that query for live MRR and extracts the Stripe computation out of
routes/admin.ts into services/mrr.ts, so the email and the dashboard read
from one implementation and cannot drift. admin.ts loses ~3200 characters
of inline logic and calls computeMrr() instead.

The report sends committed revenue (external minus already-cancelling) plus
the at-risk figure separately, so the headline is the money that will still
be there next month rather than the money currently billing.

computeMrr is fetched after the D1 batch rather than inside it: it is an
HTTP call to Stripe, not a query, and a Stripe outage must not fail the
whole report. It returns zeros on failure and callers treat a zero
subscription count as "unavailable".
